fix: Lo siento Brian, estaba super chulo lo que has hecho pero he tenido que cambiarlo. Arreglo de la lógica detrás del SwipeItem de la lista de recetas.

+++ app/src/main/java/com/alvarobrivaro/coffee/domain/DeleteRecipeUseCase.kt
@@ -4,7 +4,7 @@ import com.alvarobrivaro.coffee.data.CoffeeAppRepository
 import javax.inject.Inject
 
 class DeleteRecipeUseCase @Inject constructor(private val repository: CoffeeAppRepository) {
-    suspend operator fun invoke(recipeId: Long) {
-        repository.deleteRecipe(recipeId)
+    suspend operator fun invoke(recipeId: Int) {
+        repository.deleteRecipe(recipeId.toLong())
     }
 }
